Get Fatality Analysis Reporting System (FARS) data with the FARS API

This function provides a convenient interface for accessing FARS data or data summaries using a range of criteria. The api parameter allows you to call one of the following functions to access DOT NHTSA’s Crash API:

get_fars_crash_list returns a list of fatal crashes that have occurred in multiple states in one or more years.

get_fars_crash_details returns a details of a fatal crash that has occurred in a state for a single year.

get_fars_crashes a list of fatal crashes by location that have occurred throughout U.S.

get_fars_summary provides a count of injury severity that have occurred throughout U.S. including count of fatalities and crashes.

get_fars_year provides one of 20 FARS data tables for a single year. Supports downloading to a CSV or JSON file.

Both get_fars_crash_list and get_fars_crashes limit the returned data to 5000 records so consider limiting the range of years requested if data exceeds that threshold.

This package also enables access to the FARS data available through the NHTSA data downloads server in a zip format. Set api to "zip" or use the get_fars_zip function to download this data.

Usage

get_fars(
  year = 2022,
  state,
  county = NULL,
  api = c("crashes", "cases", "state list", "summary count", "year dataset", "zip"),
  type = NULL,
  details = FALSE,
  geometry = FALSE,
  crs = NULL,
  cases = NULL,
  vehicles = NULL,
  format = "json",
  pr = FALSE,
  path = NULL,
  download = FALSE
)

get_fars_crashes(
  year = 2022,
  start_year,
  end_year = NULL,
  state,
  county,
  details = FALSE,
  geometry = FALSE,
  crs = NULL
)

get_fars_cases(
  year = 2022,
  state,
  cases,
  details = FALSE,
  geometry = FALSE,
  crs = NULL
)

get_fars_crash_list(
  year = 2022,
  start_year = NULL,
  end_year = NULL,
  state,
  vehicles = c(1, 50)
)

get_fars_summary(year = 2022, start_year, end_year = NULL, state)

get_fars_year(
  year = 2022,
  type = "accident",
  state,
  format = "json",
  path = NULL,
  geometry = FALSE,
  crs = NULL,
  download = FALSE
)

Arguments

year

numeric vector. Year or range with start and end year. If api is "details", "year dataset", or "zip" (or using the get_fars_crash_details, get_fars_year, or get_fars_zip functions), a single year is required. All other api options support a range with the minimum value is used as a start year and the maximum value used as a end year. Most api options support the years from 2010 through the most recent year of release. "year dataset" only supports 2010 to 2017 and "zip" supports 1975 to 2022. start_year and end_year are ignored if year is not NULL.

state

Required. State name, abbreviation, or FIPS number. get_fars_crash_list supports multiple states.

county

County name or FIPS number. Required for get_fars_crashes.

api

character. API function to use. Supported values include "crashes", "cases", "state list", "summary count", "year dataset", and "zip". Default: "crashes".

type

Name of the dataset or data file to download when using the "year dataset" api or get_fars_year. Supported values include "ACCIDENT", "CEVENT", "DAMAGE", "DISTRACT", "DRIMPAIR", "FACTOR", "MANEUVER", "NMCRASH", "NMIMPAIR", "NMPRIOR", "PARKWORK", "PBTYPE", "PERSON", "SAFETYEQ", "VEHICLE", "VEVENT VINDECODE", "VINDERIVED", "VIOLATION", "VISION", and "VSOE". Lowercase or mixed case values are permitted.

details

Type of detailed crash data to return (either "events" or "vehicles"). If TRUE for get_fars or get_fars_crashes, detailed case data (excluding event and vehicle data) is attached to the returned crash data. If NULL for get_fars_cases, events and vehicle data are excluded from the returned case data. returned by get_fars_cases. Optional for get_fars_crash_details. Default: NULL for get_fars_cases; FALSE for get_fars and get_fars_crashes.

geometry

If TRUE, return sf object. Optional for get_fars_crashes.

crs

Coordinate reference system to return for get_fars_crashes if geometry is TRUE.

cases

One or more FARS case numbers. Required if api is "cases" or using get_fars_cases. Multiple case numbers can be provided.

vehicles

numeric vector with the minimum and maximum number of vehicles, e.g. c(1, 2) for minimum of 1 vehicle and maximum of 2. Required for get_fars_crash_list.

format

Default "json". "csv" is supported when using the "year dataset" api. "sas" is supporting for the "zip" api.

pr

logical. If TRUE, download zip file with FARS data for Puerto Rico. No Puerto Rico data available for years 1975-1977. Default: FALSE for get_fars_zip only.

path

File path used if download is TRUE.

download

logical. If TRUE and the api is "year dataset" or "zip", download the data to a file. Default FALSE.

start_year

Start year for crash reports.

end_year

End year for crash reports.

Examples

head(get_fars_crashes(state = "MD", county = "Baltimore city"), 5)
#>   CITY       CITYNAME COUNTY             COUNTYNAME CaseYear FATALS    LATITUDE
#> 1   50      BALTIMORE    510 BALTIMORE (CITY) (510)     2022      1 39.28737222
#> 2   50      BALTIMORE    510 BALTIMORE (CITY) (510)     2022      1 39.28748333
#> 3    0 NOT APPLICABLE    510 BALTIMORE (CITY) (510)     2022      1 39.28098611
#> 4   50      BALTIMORE    510 BALTIMORE (CITY) (510)     2022      1 39.27735833
#> 5   50      BALTIMORE    510 BALTIMORE (CITY) (510)     2022      1 39.27402222
#>        LONGITUD STATE STATENAME ST_CASE TOTALVEHICLES TWAY_ID TWAY_ID2 VE_FORMS
#> 1 -76.553625000    24  Maryland  240061             1   I-895                 1
#> 2 -76.546555560    24  Maryland  240071             1  SR-150      UU0        1
#> 3 -76.622394440    24  Maryland  240072             1  I-395A                 1
#> 4 -76.619722220    24  Maryland  240103             4   I-395                 4
#> 5 -76.621216670    24  Maryland  240144             2   I-395                 2

get_fars_cases(state = "MD", cases = "240274")
#>   ARR_HOUR                   ARR_HOURNAME ARR_MIN
#> 1       99 Unknown EMS Scene Arrival Hour      99
#>                         ARR_MINNAME CF1 CF1NAME CF2 CF2NAME CF3 CF3NAME CITY
#> 1 Unknown EMS Scene Arrival Minutes  NA      NA  NA      NA  NA      NA    0
#>         CITYNAME COUNTY       COUNTYNAME CaseYear
#> 1 NOT APPLICABLE      3 ANNE ARUNDEL (3)     2022
#>                                        CrashRFs DAY DAY_WEEK DAY_WEEKNAME
#> 1 0, None Noted, 2022, 24, MD, Maryland, 240274  19        6       Friday
#>   DRUNK_DR FATALS FUNC_SYS                                        FUNC_SYSNAME
#> 1       NA      1        2 Principal Arterial - Other Freeways and Expressways
#>   HARM_EV  HARM_EVNAME HOSP_HR                       HOSP_HRNAME HOSP_MN
#> 1       9 Pedalcyclist      88 Not Applicable (Not Transported)       88
#>                        HOSP_MNNAME HOUR      HOURNAME    LATITUDE LATITUDENAME
#> 1 Not Applicable (Not Transported)   18 6:00pm-6:59pm 39.17756667  39.17756667
#>   LGT_COND LGT_CONDNAME      LONGITUD LONGITUDNAME MAN_COLL
#> 1        1     Daylight -76.605758330 -76.60575833        0
#>                                                                    MAN_COLLNAME
#> 1 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#>   MILEPT MILEPTNAME MINUTE MINUTENAME MONTH MonthName NHS
#> 1     50         50     20         20     8    August   1
#>                      NHSNAME
#> 1 This section IS ON the NHS
#>                                                                                    NMDrugs
#> 1 2022, 1, Tested, No Drugs Found/Negative, 1, Whole Blood, 1, 24, MD, Maryland, 240274, 0
#>                                            NMPersonRF
#> 1 2022, 0, None Noted, 1, 24, MD, Maryland, 240274, 0
#>                                                        NMRace NOT_HOUR
#> 1 2022, 0, No, 1, 1, 99, Unknown, 24, MD, Maryland, 240274, 0       99
#>   NOT_HOURNAME NOT_MIN NOT_MINNAME
#> 1      Unknown      99     Unknown
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  NPersons
#> 1 61, 61 Years, 97, Not a Motor Vehicle Occupant, NA, NA, 0, 0.000 % BAC, 2, Test Given, 1, Blood, NA, NA, NA, 3, ANNE ARUNDEL (3), 2022, 19, 19, 19, 18, 18:00-18:59, 30, 30, 8, August, 1830, 1830, 2022, 2022, 3, Unknown/Not Reported if Motorized, 3, Bicycle, 7, Died at Scene, 0, No (Alcohol Not Involved), NA, NA, NA, NA, NA, NA, 0, No (drugs not involved), NA, NA, NA, NA, NA, NA, NA, NA, 2, Test Given, 8, Not Applicable, 0, Ejection Path Not Applicable, NA, NA, 0, Not Extricated or Not Applicable, NA, NA, 2, Principal Arterial - Other Freeways and Expressways, NA, NA, NA, NA, 9, Pedalcyclist, 8, Not a Motor Vehicle Occupant, 96, Not a Motor Vehicle Occupant, 99, Unknown, 0, Not Transported for Treatment, 18, 6:00pm-6:59pm, NA, NA, NA, NA, NA, NA, 4, Fatal Injury (K), 0, 0, 10, 10, 11, Not at Intersection - On Roadway, Not in Marked Crosswalk, NA, NA, NA, NA, 0, The First Harmful Event was Not a Collision with a Motor Vehicle in Transport, 20, 20, NA, NA, 8, August, NA, NA, 1, 6, Bicyclist, NA, NA, NA, NA, NA, NA, NA, NA, NA, 8, Not a Motor Vehicle Occupant, 96, Not a Motor Vehicle Occupant, NA, NA, NA, NA, 2, Urban, NA, 0, No, 0, Not a Motor Vehicle Occupant, 1, Male, NA, NA, 24, Maryland, 1, 01, 240274, NA, NA, 0, 1, NA, NA, NA, NA, NA, NA, 9, Unknown
#>                                                                                                                                                                                NmCrashes
#> 1 2022, 2022, 2, 7, Failure to Yield Right-Of-Way, Improper Turn/Merge, 2, 7, Failure to Yield Right-Of-Way, Improper Turn/Merge, 1, 1, 24, 24, Maryland, Maryland, 240274, 240274, 0, 0
#>                                                                 NmDistract
#> 1 2022, 96, Not Reported, 96, Not Reported, 1, 24, MD, Maryland, 240274, 0
#>                                                               NmImpairs
#> 1 2022, 99, Reported as Unknown if Impaired, 1, 24, Maryland, 240274, 0
#>                                                                     NmPriors
#> 1 2022, 3, Crossing Roadway, 3, Crossing Roadway, 1, 24, Maryland, 240274, 0
#>   PEDS PERMVIT PERNOTMVIT PERSONS PVH_INVL ParkWorks
#> 1    1       1          1       1        0        NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                  PbTypes
#> 1 220, Bicyclist Left Turn / Merge, 221, Bicyclist Left Turn - Same Direction, 1, With Traffic, 3, Not At Intersection, 1, Travel Lane, 2022, 7, Not a Pedestrian, 7, Not a Pedestrian, 61, 61 Years, 0, None Noted, 6, Bicyclist, 1, Male, 0, None Noted, 0, None Noted, 0, Not a Pedestrian, 0, Not a Pedestrian, 7, Not a Pedestrian, 7, Not a Pedestrian, 7, Not a Pedestrian, 77, Not a Pedestrian, 7, Not a Pedestrian, 1, 24, Maryland, 240274, 0
#>      RAIL       RAILNAME RD_OWNER         RD_OWNERNAME RELJCT1 RELJCT1NAME
#> 1 0000000 Not Applicable        1 State Highway Agency       0          No
#>   RELJCT2  RELJCT2NAME REL_ROAD REL_ROADNAME ROAD_FNC ROAD_FNCNAME ROUTE
#> 1       1 Non-Junction        1   On Roadway       NA           NA     3
#>       ROUTENAME RUR_URB RUR_URBNAME SCH_BUS SCH_BUSNAME SP_JUR
#> 1 State Highway       2       Urban       0          No      0
#>                SP_JURNAME STATENAME ST_CASE
#> 1 No Special Jurisdiction  Maryland  240274
#>                                  SafetyEQs State TWAY_ID TWAY_ID2 TYP_INT
#> 1 2022, NA, NA, 1, 24, Maryland, 240274, 0    24   SR-10                1
#>           TYP_INTNAME VE_FORMS VE_TOTAL WEATHER WEATHER1 WEATHER1NAME WEATHER2
#> 1 Not an Intersection        1        1       1       NA           NA       NA
#>   WEATHER2NAME WEATHERNAME WRK_ZONE WRK_ZONENAME
#> 1           NA       Clear        0         None
#>                                   Weathers YEAR
#> 1 2022, 24, MD, Maryland, 240274, 1, Clear 2022

get_fars_crash_list(state = "MD", vehicles = 5)
#>               CountyName                  CrashDate Fatals Peds Persons St_Case
#> 1       ANNE ARUNDEL (3) /Date(1652041560000-0400)/      1    0      10  240139
#> 2   PRINCE GEORGE`S (33) /Date(1657938900000-0400)/      1    0       2  240236
#> 3       ANNE ARUNDEL (3) /Date(1663643520000-0400)/      1    0       5  240332
#> 4 BALTIMORE (CITY) (510) /Date(1656433320000-0400)/      1    0      11  240394
#> 5       ANNE ARUNDEL (3) /Date(1668836580000-0500)/      1    0       7  240423
#> 6           HARFORD (25) /Date(1670413440000-0500)/      1    0       5  240435
#> 7   PRINCE GEORGE`S (33) /Date(1644579120000-0500)/      1    0       6  240501
#>   State StateName TotalVehicles
#> 1    24  Maryland             5
#> 2    24  Maryland             5
#> 3    24  Maryland             5
#> 4    24  Maryland             5
#> 5    24  Maryland             5
#> 6    24  Maryland             5
#> 7    24  Maryland             5

get_fars_summary(state = "MD")
#>   CaseYear CrashCounts TotalFatalCounts
#> 1     2022         532              564

head(get_fars_year(state = "MD", type = "PERSON"), 5)
#>   AGE  AGENAME AIR_BAG           AIR_BAGNAME ALC_DET ALC_DETNAME ALC_RES
#> 1  93 93 Years      20          Not Deployed      NA          NA       0
#> 2  29 29 Years       1       Deployed- Front      NA          NA     996
#> 3  33 33 Years       8 Deployed- Combination      NA          NA     996
#> 4  13 13 Years       8 Deployed- Combination      NA          NA       0
#> 5  32 32 Years       8 Deployed- Combination      NA          NA     996
#>      ALC_RESNAME ALC_STATUS ALC_STATUSNAME ATST_TYP    ATST_TYPNAME BODY_TYP
#> 1    0.000 % BAC          2     Test Given        1           Blood        4
#> 2 Test Not Given          0 Test Not Given        0  Test Not Given       14
#> 3 Test Not Given          0 Test Not Given        0  Test Not Given        4
#> 4    0.000 % BAC          2     Test Given        8 Other Test Type        4
#> 5 Test Not Given          0 Test Not Given        0  Test Not Given        4
#>                                                         BODY_TYPNAME CASEYEAR
#> 1                                              4-door sedan, hardtop     2022
#> 2 Compact Utility (Utility Vehicle Categories "Small" and "Midsize")     2022
#> 3                                              4-door sedan, hardtop     2022
#> 4                                              4-door sedan, hardtop     2022
#> 5                                              4-door sedan, hardtop     2022
#>   CERT_NO COUNTY      COUNTYNAME DAY DEATH_DA               DEATH_DANAME
#> 1      NA     43 WASHINGTON (43)   1        1                          1
#> 2      NA     43 WASHINGTON (43)   1       88 Not Applicable (Non-Fatal)
#> 3      NA      5   BALTIMORE (5)   1       88 Not Applicable (Non-Fatal)
#> 4      NA      5   BALTIMORE (5)   1        1                          1
#> 5      NA      5   BALTIMORE (5)   1       88 Not Applicable (Non-Fatal)
#>   DEATH_HR               DEATH_HRNAME DEATH_MN               DEATH_MNNAME
#> 1       11                11:00-11:59        5                          5
#> 2       88 Not Applicable (Non-fatal)       88 Not Applicable (Non-fatal)
#> 3       88 Not Applicable (Non-fatal)       88 Not Applicable (Non-fatal)
#> 4       22                22:00-22:59       16                         16
#> 5       88 Not Applicable (Non-fatal)       88 Not Applicable (Non-fatal)
#>   DEATH_MO               DEATH_MONAME DEATH_TM               DEATH_TMNAME
#> 1        1                    January     1105                       1105
#> 2       88 Not Applicable (Non-Fatal)     8888 Not Applicable (Non-fatal)
#> 3       88 Not Applicable (Non-Fatal)     8888 Not Applicable (Non-fatal)
#> 4        1                    January     2216                       2216
#> 5       88 Not Applicable (Non-Fatal)     8888 Not Applicable (Non-fatal)
#>   DEATH_YR               DEATH_YRNAME DEVMOTOR DEVMOTORNAME DEVTYPE DEVTYPENAME
#> 1     2022                       2022     <NA>         <NA>    <NA>        <NA>
#> 2     8888 Not Applicable (Non-fatal)     <NA>         <NA>    <NA>        <NA>
#> 3     8888 Not Applicable (Non-fatal)     <NA>         <NA>    <NA>        <NA>
#> 4     2022                       2022     <NA>         <NA>    <NA>        <NA>
#> 5     8888 Not Applicable (Non-fatal)     <NA>         <NA>    <NA>        <NA>
#>   DOA        DOANAME DRINKING              DRINKINGNAME DRUGRES1 DRUGRES1NAME
#> 1   0 Not Applicable        8              Not Reported       NA           NA
#> 2   0 Not Applicable        0 No (Alcohol Not Involved)       NA           NA
#> 3   0 Not Applicable        0 No (Alcohol Not Involved)       NA           NA
#> 4   0 Not Applicable        8              Not Reported       NA           NA
#> 5   0 Not Applicable        8              Not Reported       NA           NA
#>   DRUGRES2 DRUGRES2NAME DRUGRES3 DRUGRES3NAME DRUGS               DRUGSNAME
#> 1       NA           NA       NA           NA     8            Not Reported
#> 2       NA           NA       NA           NA     0 No (drugs not involved)
#> 3       NA           NA       NA           NA     0 No (drugs not involved)
#> 4       NA           NA       NA           NA     8            Not Reported
#> 5       NA           NA       NA           NA     8            Not Reported
#>   DRUGTST1 DRUGTST1NAME DRUGTST2 DRUGTST2NAME DRUGTST3 DRUGTST3NAME DRUG_DET
#> 1       NA           NA       NA           NA       NA           NA       NA
#> 2       NA           NA       NA           NA       NA           NA       NA
#> 3       NA           NA       NA           NA       NA           NA       NA
#> 4       NA           NA       NA           NA       NA           NA       NA
#> 5       NA           NA       NA           NA       NA           NA       NA
#>   DRUG_DETNAME DSTATUS    DSTATUSNAME EJECTION    EJECTIONNAME EJ_PATH
#> 1           NA       2     Test Given        1 Totally Ejected       9
#> 2           NA       0 Test Not Given        0     Not Ejected       0
#> 3           NA       0 Test Not Given        0     Not Ejected       0
#> 4           NA       2     Test Given        0     Not Ejected       0
#> 5           NA       0 Test Not Given        0     Not Ejected       0
#>                    EJ_PATHNAME EMER_USE   EMER_USENAME EXTRICAT
#> 1        Ejection Path Unknown        0 Not Applicable        0
#> 2 Ejection Path Not Applicable        0 Not Applicable        0
#> 3 Ejection Path Not Applicable        0 Not Applicable        0
#> 4 Ejection Path Not Applicable        0 Not Applicable        0
#> 5 Ejection Path Not Applicable        0 Not Applicable        0
#>                       EXTRICATNAME FIRE_EXP       FIRE_EXPNAME FUNC_SYS
#> 1 Not Extricated or Not Applicable        0 No or Not Reported        3
#> 2 Not Extricated or Not Applicable        0 No or Not Reported        3
#> 3 Not Extricated or Not Applicable        0 No or Not Reported        1
#> 4 Not Extricated or Not Applicable        0 No or Not Reported        1
#> 5 Not Extricated or Not Applicable        0 No or Not Reported        1
#>                 FUNC_SYSNAME GVWR_FROM
#> 1 Principal Arterial - Other        11
#> 2 Principal Arterial - Other        11
#> 3                 Interstate        11
#> 4                 Interstate        11
#> 5                 Interstate        11
#>                                    GVWR_FROMNAME GVWR_TO
#> 1 Class 1: 6,000 lbs. or less (2,722 kg or less)      12
#> 2 Class 1: 6,000 lbs. or less (2,722 kg or less)      11
#> 3 Class 1: 6,000 lbs. or less (2,722 kg or less)      11
#> 4 Class 1: 6,000 lbs. or less (2,722 kg or less)      11
#> 5 Class 1: 6,000 lbs. or less (2,722 kg or less)      11
#>                                       GVWR_TONAME HARM_EV
#> 1 Class 2: 6,001 - 10,000 lbs. (2,722 - 4,536 kg)      12
#> 2  Class 1: 6,000 lbs. or less (2,722 kg or less)      12
#> 3  Class 1: 6,000 lbs. or less (2,722 kg or less)      25
#> 4  Class 1: 6,000 lbs. or less (2,722 kg or less)      25
#> 5  Class 1: 6,000 lbs. or less (2,722 kg or less)      25
#>                  HARM_EVNAME HELM_MIS             HELM_MISNAME HELM_USE
#> 1 Motor Vehicle In-Transport        7 None Used/Not Applicable       20
#> 2 Motor Vehicle In-Transport        7 None Used/Not Applicable       20
#> 3   Concrete Traffic Barrier        7 None Used/Not Applicable       20
#> 4   Concrete Traffic Barrier        7 None Used/Not Applicable       20
#> 5   Concrete Traffic Barrier        7 None Used/Not Applicable       20
#>     HELM_USENAME HISPANIC                    HISPANICNAME HOSPITAL
#> 1 Not Applicable       99                         Unknown        0
#> 2 Not Applicable        0 Not A Fatality (not Applicable)        5
#> 3 Not Applicable        0 Not A Fatality (not Applicable)        0
#> 4 Not Applicable       99                         Unknown        5
#> 5 Not Applicable        0 Not A Fatality (not Applicable)        0
#>                    HOSPITALNAME HOUR      HOURNAME ICFINALBODY ICFINALBODYNAME
#> 1 Not Transported for Treatment    9 9:00am-9:59am           0  Not Applicable
#> 2                    EMS Ground    9 9:00am-9:59am           0  Not Applicable
#> 3 Not Transported for Treatment   21 9:00pm-9:59pm           0  Not Applicable
#> 4                    EMS Ground   21 9:00pm-9:59pm           0  Not Applicable
#> 5 Not Transported for Treatment   21 9:00pm-9:59pm           0  Not Applicable
#>   IMPACT1    IMPACT1NAME IMPACT2 IMPACT2NAME INJ_SEV                INJ_SEVNAME
#> 1      12 12 Clock Point      NA          NA       4           Fatal Injury (K)
#> 2      12 12 Clock Point      NA          NA       2 Suspected Minor Injury (B)
#> 3      12 12 Clock Point      NA          NA       1        Possible Injury (C)
#> 4      12 12 Clock Point      NA          NA       4           Fatal Injury (K)
#> 5      12 12 Clock Point      NA          NA       1        Possible Injury (C)
#>   LAG_HRS LAG_HRSNAME LAG_MINS LAG_MINSNAME LOCATION
#> 1       1           1       17           17        0
#> 2     999     Unknown       99      Unknown        0
#> 3     999     Unknown       99      Unknown        0
#> 4       1           1       10           10        0
#> 5     999     Unknown       99      Unknown        0
#>                  LOCATIONNAME MAKE      MAKENAME MAK_MOD
#> 1 Occupant of a Motor Vehicle   42 Mercedes-Benz   42031
#> 2 Occupant of a Motor Vehicle   35 Nissan/Datsun   35401
#> 3 Occupant of a Motor Vehicle   55       Hyundai   55035
#> 4 Occupant of a Motor Vehicle   55       Hyundai   55035
#> 5 Occupant of a Motor Vehicle   55       Hyundai   55035
#>                                               MAK_MODNAME MAN_COLL
#> 1 Mercedes-Benz 200/220/230/240/ 250/260/280/300/ 320/420        2
#> 2                                Nissan/Datsun Pathfinder        2
#> 3                                         Hyundai Elantra        0
#> 4                                         Hyundai Elantra        0
#> 5                                         Hyundai Elantra        0
#>                                                                    MAN_COLLNAME
#> 1                                                                Front-to-Front
#> 2                                                                Front-to-Front
#> 3 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#> 4 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#> 5 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#>   MINUTE MINUTENAME MOD_YEAR MOD_YEARNAME MONTH MONTHNAME N_MOT_NO PER_NO
#> 1     48         48     1988         1988     1   January       NA      1
#> 2     48         48     2002         2002     1   January       NA      1
#> 3      6          6     2021         2021     1   January       NA      1
#> 4      6          6     2021         2021     1   January       NA      2
#> 5      6          6     2021         2021     1   January       NA      3
#>   PER_TYP                               PER_TYPNAME P_SF1 P_SF1NAME P_SF2
#> 1       1    Driver of a Motor Vehicle In-Transport    NA        NA    NA
#> 2       1    Driver of a Motor Vehicle In-Transport    NA        NA    NA
#> 3       1    Driver of a Motor Vehicle In-Transport    NA        NA    NA
#> 4       2 Passenger of a Motor Vehicle In-Transport    NA        NA    NA
#> 5       2 Passenger of a Motor Vehicle In-Transport    NA        NA    NA
#>   P_SF2NAME P_SF3 P_SF3NAME RACE RACENAME REST_MIS             REST_MISNAME
#> 1        NA    NA        NA   NA       NA        7 None Used/Not Applicable
#> 2        NA    NA        NA   NA       NA        0  No Indication of Misuse
#> 3        NA    NA        NA   NA       NA        0  No Indication of Misuse
#> 4        NA    NA        NA   NA       NA        0  No Indication of Misuse
#> 5        NA    NA        NA   NA       NA        0  No Indication of Misuse
#>   REST_USE               REST_USENAME ROAD_FNC ROAD_FNCNAME ROLLOVER
#> 1       20   None Used/Not Applicable       NA           NA        0
#> 2        3 Shoulder and Lap Belt Used       NA           NA        0
#> 3        3 Shoulder and Lap Belt Used       NA           NA        0
#> 4        3 Shoulder and Lap Belt Used       NA           NA        0
#> 5        3 Shoulder and Lap Belt Used       NA           NA        0
#>   ROLLOVERNAME RUR_URB RUR_URBNAME SCH_BUS SCH_BUSNAME SEAT_POS
#> 1  No Rollover       2       Urban       0          No       11
#> 2  No Rollover       2       Urban       0          No       11
#> 3  No Rollover       2       Urban       0          No       11
#> 4  No Rollover       2       Urban       0          No       23
#> 5  No Rollover       2       Urban       0          No       13
#>              SEAT_POSNAME SEX SEXNAME SPEC_USE         SPEC_USENAME STATE
#> 1   Front Seat, Left Side   1    Male        0 No Special Use Noted    24
#> 2   Front Seat, Left Side   1    Male        0 No Special Use Noted    24
#> 3   Front Seat, Left Side   1    Male        0 No Special Use Noted    24
#> 4 Second Seat, Right Side   1    Male        0 No Special Use Noted    24
#> 5  Front Seat, Right Side   2  Female        0 No Special Use Noted    24
#>   STATENAME STR_VEH STR_VEHNAME ST_CASE TOW_VEH TOW_VEHNAME VEH_NO VE_FORMS
#> 1  Maryland       0           0  240001       0 No Trailers      1        2
#> 2  Maryland       0           0  240001       0 No Trailers      2        2
#> 3  Maryland       0           0  240002       0 No Trailers      1        2
#> 4  Maryland       0           0  240002       0 No Trailers      1        2
#> 5  Maryland       0           0  240002       0 No Trailers      1        2
#>   VPICBODYCLASS                                       VPICBODYCLASSNAME
#> 1            13                                            Sedan/Saloon
#> 2             7 Sport Utility Vehicle (SUV)/Multi-Purpose Vehicle (MPV)
#> 3            13                                            Sedan/Saloon
#> 4            13                                            Sedan/Saloon
#> 5            13                                            Sedan/Saloon
#>   VPICMAKE  VPICMAKENAME VPICMODEL VPICMODELNAME WORK_INJ
#> 1      449 Mercedes-Benz     14014           300        9
#> 2      478        Nissan      1911    Pathfinder        8
#> 3      498       Hyundai      2735       Elantra        8
#> 4      498       Hyundai      2735       Elantra        9
#> 5      498       Hyundai      2735       Elantra        8
#>                      WORK_INJNAME
#> 1                         Unknown
#> 2 Not Applicable (not a fatality)
#> 3 Not Applicable (not a fatality)
#> 4                         Unknown
#> 5 Not Applicable (not a fatality)