Get Fatality Analysis Reporting System (FARS) data with the FARS API

This function provides a convenient interface for accessing FARS data or data summaries using a range of criteria. The api parameter allows you to call one of the following functions to access DOT NHTSA’s Crash API:

get_fars_crash_list returns a list of fatal crashes that have occurred in multiple states in one or more years.

get_fars_crash_details returns a details of a fatal crash that has occurred in a state for a single year.

get_fars_crashes a list of fatal crashes by location that have occurred throughout U.S.

get_fars_summary provides a count of injury severity that have occurred throughout U.S. including count of fatalities and crashes.

get_fars_year provides one of 20 FARS data tables for a single year. Supports downloading to a CSV or JSON file.

Both get_fars_crash_list and get_fars_crashes limit the returned data to 5000 records so consider limiting the range of years requested if data exceeds that threshold.

This package also enables access to the FARS data available through the NHTSA data downloads server in a zip format. Set api to "zip" or use the get_fars_zip function to download this data.

Usage

get_fars(
  year = 2020,
  state,
  county = NULL,
  api = c("crashes", "cases", "state list", "summary count", "year dataset", "zip"),
  type,
  details = FALSE,
  geometry = FALSE,
  crs = NULL,
  cases,
  vehicles,
  format = "json",
  pr,
  path = NULL,
  download = FALSE
)

get_fars_crashes(
  year = 2020,
  start_year,
  end_year = NULL,
  state,
  county,
  details = FALSE,
  geometry = FALSE,
  crs = NULL
)

get_fars_cases(
  year = 2020,
  state,
  cases,
  details = NULL,
  geometry = FALSE,
  crs = NULL
)

get_fars_crash_list(
  year = 2020,
  start_year,
  end_year = NULL,
  state,
  vehicles = c(1, 50)
)

get_fars_summary(year = 2020, start_year, end_year = NULL, state)

get_fars_year(
  year = 2020,
  type = "accident",
  state,
  format = "json",
  path = NULL,
  geometry = FALSE,
  crs = NULL,
  download = FALSE
)

Arguments

year

numeric vector. Year or range with start and end year. If api is "details", "year dataset", or "zip" (or using the get_fars_crash_details, get_fars_year, or get_fars_zip functions), a single year is required. All other api options support a range with the minimum value is used as a start year and the maximum value used as a end year. Most api options support the years from 2010 through the most recent year of release. "year dataset" only supports 2010 to 2017 and "zip" supports 1975 to 2020. start_year and end_year are ignored if year is not NULL.

state

Required. State name, abbreviation, or FIPS number. get_fars_crash_list supports multiple states.

county

County name or FIPS number. Required for get_fars_crashes.

api

character. API function to use. Supported values include "crashes", "cases", "state list", "summary count", "year dataset", and "zip". Default: "crashes".

type

Name of the dataset or data file to download when using the "year dataset" api or get_fars_year. Supported values include "ACCIDENT", "CEVENT", "DAMAGE", "DISTRACT", "DRIMPAIR", "FACTOR", "MANEUVER", "NMCRASH", "NMIMPAIR", "NMPRIOR", "PARKWORK", "PBTYPE", "PERSON", "SAFETYEQ", "VEHICLE", "VEVENT VINDECODE", "VINDERIVED", "VIOLATION", "VISION", and "VSOE". Lowercase or mixed case values are permitted.

details

Type of detailed crash data to return (either "events" or "vehicles"). If TRUE for get_fars or get_fars_crashes, detailed case data (excluding event and vehicle data) is attached to the returned crash data. If NULL for get_fars_cases, events and vehicle data are excluded from the returned case data. returned by get_fars_cases. Optional for get_fars_crash_details. Default: NULL for get_fars_cases; FALSE for get_fars and get_fars_crashes.

geometry

If TRUE, return sf object. Optional for get_fars_crashes.

crs

Coordinate reference system to return for get_fars_crashes if geometry is TRUE.

cases

One or more FARS case numbers. Required if api is "cases" or using get_fars_cases. Multiple case numbers can be provided.

vehicles

numeric vector with the minimum and maximum number of vehicles, e.g. c(1, 2) for minimum of 1 vehicle and maximum of 2. Required for get_fars_crash_list.

format

Default "json". "csv" is supported when using the "year dataset" api. "sas" is supporting for the "zip" api.

pr

logical. If TRUE, download zip file with FARS data for Puerto Rico. No Puerto Rico data available for years 1975-1977. Default: FALSE for get_fars_zip only.

path

File path used if download is TRUE.

download

logical. If TRUE and the api is "year dataset" or "zip", download the data to a file. Default FALSE.

start_year

Start year for crash reports.

end_year

End year for crash reports.

Examples

head(get_fars_crashes(state = "MD", county = "Baltimore city"), 5)
#>   city  cityname county      countyname case_year fatals latitude  longitud
#> 1   50 BALTIMORE    510 BALTIMORE (510)      2020      1 39.29745 -76.59408
#> 2   50 BALTIMORE    510 BALTIMORE (510)      2020      1 39.26703 -76.60339
#> 3   50 BALTIMORE    510 BALTIMORE (510)      2020      1 39.30192 -76.54154
#> 4   50 BALTIMORE    510 BALTIMORE (510)      2020      1 39.29782 -76.60842
#> 5   50 BALTIMORE    510 BALTIMORE (510)      2020      1 39.32635 -76.67879
#>   state statename st_case totalvehicles tway_id tway_id2 ve_forms
#> 1    24  Maryland  240070             1 MU-1445     <NA>        1
#> 2    24  Maryland  240089             2    I-95     <NA>        2
#> 3    24  Maryland  240129             2  SR-151     <NA>        2
#> 4    24  Maryland  240141             1 MU-1370     <NA>        1
#> 5    24  Maryland  240160             3   SR-26  MU-2900        3

get_fars_cases(state = "MD", cases = "240274")
#>   ARR_HOUR                   ARR_HOURNAME ARR_MIN
#> 1       99 Unknown EMS Scene Arrival Hour      99
#>                         ARR_MINNAME CF1 CF1NAME CF2 CF2NAME CF3 CF3NAME CITY
#> 1 Unknown EMS Scene Arrival Minutes  NA      NA  NA      NA  NA      NA   50
#>    CITYNAME COUNTY      COUNTYNAME CaseYear
#> 1 BALTIMORE    510 BALTIMORE (510)     2020
#>                                  CrashRFs DAY DAY_WEEK DAY_WEEKNAME DRUNK_DR
#> 1 0, None, 2020, 24, MD, Maryland, 240274  10        3      Tuesday        0
#>   FATALS FUNC_SYS               FUNC_SYSNAME HARM_EV HARM_EVNAME HOSP_HR
#> 1      1        3 Principal Arterial - Other       8  Pedestrian      99
#>   HOSP_HRNAME HOSP_MN                       HOSP_MNNAME HOUR        HOURNAME
#> 1     Unknown      99 Unknown EMS Hospital Arrival Time   23 11:00pm-11:59pm
#>      LATITUDE LATITUDENAME LGT_COND       LGT_CONDNAME      LONGITUD
#> 1 39.29421111  39.29421111        2 Dark - Not Lighted -76.668097220
#>    LONGITUDNAME MAN_COLL
#> 1 -76.668097220        0
#>                                                                    MAN_COLLNAME
#> 1 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#>   MILEPT MILEPTNAME MINUTE MINUTENAME MONTH MonthName NHS
#> 1     23         23      0          0     3     March   1
#>                      NHSNAME
#> 1 This section IS ON the NHS
#>                                                                                                                                                                                                                                                                                                                                                                                                   NMDrugs
#> 1 2020, 2020, 2020, 2020, 2020, 2020, 407, 167, 151, 996, 151, 167, COCAINE, METHADONE, FENTANYL, Other Drug, FENTANYL, METHADONE, 2, 2, 2, 2, 1, 1, Urine, Urine, Urine, Urine, Whole Blood, Whole Blood, 1, 1, 1, 1, 1, 1, 24, 24, 24, 24, 24, 24, MD, MD, MD, MD, MD, MD, Maryland, Maryland, Maryland, Maryland, Maryland, Maryland, 240274, 240274, 240274, 240274, 240274, 240274, 0, 0, 0, 0, 0, 0
#>                                      NMPersonRF
#> 1 2020, 0, None, 1, 24, MD, Maryland, 240274, 0
#>                                                                         NMRace
#> 1 2020, False, 1, 1, 2, Black or African American, 24, MD, Maryland, 240274, 0
#>   NOT_HOUR NOT_HOURNAME NOT_MIN NOT_MINNAME
#> 1       99      Unknown      99     Unknown
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    NPersons
#> 1 51, 51 Years, 97, Not a Motor Vehicle Occupant, 9, Not Reported, 0, 0.000 % BAC, 2, Test Given, 1, Blood, NA, NA, NA, 510, BALTIMORE (510), 2020, 10, 10, 10, 23, 23:00-23:59, 59, 59, 3, March, 2359, 2359, 2020, 2020, 0, Not Applicable, 0, No (Alcohol Not Involved), NA, NA, NA, NA, NA, NA, 0, No (drugs not involved), NA, NA, NA, NA, NA, NA, 8, Not Reported, 2, Test Given, 8, Not Applicable, 0, Ejection Path Not Applicable, NA, NA, 0, Not Extricated or Not Applicable, NA, NA, 3, Principal Arterial - Other, 8, Pedestrian, 8, Not a Motor Vehicle Occupant, 96, Not a Motor Vehicle Occupant, 7, Non-Hispanic, 5, EMS Ground, 23, 11:00pm-11:59pm, NA, NA, NA, NA, NA, NA, 4, Fatal Injury (K), 0, 0, 59, 59, 11, Not at Intersection - On Roadway, Not in Marked Crosswalk, NA, NA, NA, NA, 0, The First Harmful Event was Not a Collision with a Motor Vehicle in Transport, 0, 0, NA, NA, 3, March, NA, NA, 1, 5, Pedestrian, NA, NA, NA, NA, NA, NA, NA, NA, NA, 8, Not a Motor Vehicle Occupant, 96, Not a Motor Vehicle Occupant, NA, NA, NA, NA, 2, Urban, NA, 0, No, 0, Not a Motor Vehicle Occupant, 1, Male, NA, NA, 24, Maryland, 1, NA, 240274, NA, NA, 0, 1, NA, NA, NA, NA, NA, NA, 0, No
#>                                                                    NmCrashes
#> 1 2020, NA, NA, 2, Failure to Yield Right-Of-Way, 1, 24, Maryland, 240274, 0
#>                                                        NmDistract
#> 1 2020, NA, NA, 0, Not Distracted, 1, 24, MD, Maryland, 240274, 0
#>                                            NmImpairs
#> 1 2020, 98, Not Reported, 1, 24, Maryland, 240274, 0
#>                                                        NmPriors PEDS PERMVIT
#> 1 2020, NA, NA, 3, Crossing Roadway, 1, 24, Maryland, 240274, 0    1       1
#>   PERNOTMVIT PERSONS PVH_INVL ParkWorks
#> 1          1       1        0        NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    PbTypes
#> 1 0, 0, Not a Cyclist, Not a Cyclist, 0, 0, Not a Cyclist, Not a Cyclist, 7, 7, Not a Cyclist, Not a Cyclist, 7, 7, Not a Cyclist, Not a Cyclist, 7, 7, Not a Cyclist, Not a Cyclist, 2020, 2020, 8, 8, Not Applicable, Not Applicable, 8, 8, Not Applicable, Not Applicable, 51, 51, 51 Years, 51 Years, 0, 0, None Noted, None Noted, 5, 5, Pedestrian, Pedestrian, 1, 1, Male, Male, 1, 1, Yes, Yes, 0, 0, None Noted, None Noted, 750, 750, Crossing Roadway - Vehicle Not Turning, Crossing Roadway - Vehicle Not Turning, 760, 760, Pedestrian Failed to Yield, Pedestrian Failed to Yield, 8, 8, Not Applicable, Not Applicable, 8, 8, Not Applicable, Not Applicable, 3, 3, Not At Intersection, Not At Intersection, 3, 3, Travel Lane, Travel Lane, 8, 8, Not Applicable, Not Applicable, 1, 1, 24, 24, Maryland, Maryland, 240274, 240274, 0, 1
#>      RAIL       RAILNAME RD_OWNER         RD_OWNERNAME RELJCT1 RELJCT1NAME
#> 1 0000000 Not Applicable        1 State Highway Agency       0          No
#>   RELJCT2  RELJCT2NAME REL_ROAD REL_ROADNAME ROAD_FNC ROAD_FNCNAME ROUTE
#> 1       1 Non-Junction        1   On Roadway       NA           NA     2
#>      ROUTENAME RUR_URB RUR_URBNAME SCH_BUS SCH_BUSNAME SP_JUR
#> 1 U.S. Highway       2       Urban       0          No      0
#>                SP_JURNAME STATENAME ST_CASE
#> 1 No Special Jurisdiction  Maryland  240274
#>                                  SafetyEQs State TWAY_ID TWAY_ID2 TYP_INT
#> 1 2020, NA, NA, 1, 24, Maryland, 240274, 0    24   US-40       NA       1
#>           TYP_INTNAME VE_FORMS VE_TOTAL WEATHER WEATHER1 WEATHER1NAME WEATHER2
#> 1 Not an Intersection        1        1       1       NA           NA       NA
#>   WEATHER2NAME WEATHERNAME WRK_ZONE            WRK_ZONENAME
#> 1           NA       Clear        4 Work Zone, Type Unknown
#>                                   Weathers YEAR
#> 1 2020, 24, MD, Maryland, 240274, 1, Clear 2020

get_fars_crash_list(state = "MD", vehicles = 5)
#>       county_name                 crash_date fatals peds persons st_case state
#> 1 MONTGOMERY (31) /Date(1599691020000-0400)/      2    0       5  240297    24
#> 2 BALTIMORE (510) /Date(1594930320000-0400)/      1    0       4  240307    24
#> 3 BALTIMORE (510) /Date(1592318340000-0400)/      1    0       7  240407    24
#> 4 ST. MARY`S (37) /Date(1607365080000-0500)/      1    0       7  240467    24
#>   state_name total_vehicles
#> 1   Maryland              5
#> 2   Maryland              5
#> 3   Maryland              5
#> 4   Maryland              5

get_fars_summary(state = "MD")
#>   case_year crash_counts total_fatal_counts
#> 1      2020          540                567

head(get_fars_year(state = "MD", type = "PERSON"), 5)
#>   AGE  AGENAME AIR_BAG                  AIR_BAGNAME ALC_DET  ALC_DETNAME
#> 1  28 28 Years      97 Not a Motor Vehicle Occupant       9 Not Reported
#> 2  37 37 Years      20                 Not Deployed       9 Not Reported
#> 3   6  6 Years      97 Not a Motor Vehicle Occupant       9 Not Reported
#> 4  38 38 Years      97 Not a Motor Vehicle Occupant       9 Not Reported
#> 5  78 78 Years      20                 Not Deployed       9 Not Reported
#>   ALC_RES    ALC_RESNAME ALC_STATUS ALC_STATUSNAME ATST_TYP   ATST_TYPNAME
#> 1      50    0.050 % BAC          2     Test Given        1          Blood
#> 2     996 Test Not Given          0 Test Not Given        0 Test Not Given
#> 3       0    0.000 % BAC          2     Test Given        1          Blood
#> 4     996 Test Not Given          0 Test Not Given        0 Test Not Given
#> 5     996 Test Not Given          0 Test Not Given        0 Test Not Given
#>   BODY_TYP               BODY_TYPNAME CASEYEAR CERT_NO COUNTY   COUNTYNAME DAY
#> 1     <NA>                       <NA>     2020      NA     13 CARROLL (13)   2
#> 2        2 2-door sedan,hardtop,coupe     2020      NA     13 CARROLL (13)   2
#> 3     <NA>                       <NA>     2020      NA     27  HOWARD (27)   6
#> 4     <NA>                       <NA>     2020      NA     27  HOWARD (27)   6
#> 5       34               Light Pickup     2020      NA     27  HOWARD (27)   6
#>   DEATH_DA               DEATH_DANAME DEATH_HR               DEATH_HRNAME
#> 1        2                          2       22                22:00-22:59
#> 2       88 Not Applicable (Non-Fatal)       88 Not Applicable (Non-fatal)
#> 3        6                          6       15                15:00-15:59
#> 4       88 Not Applicable (Non-Fatal)       88 Not Applicable (Non-fatal)
#> 5       88 Not Applicable (Non-Fatal)       88 Not Applicable (Non-fatal)
#>   DEATH_MN               DEATH_MNNAME DEATH_MO               DEATH_MONAME
#> 1       16                         16        1                    January
#> 2       88 Not Applicable (Non-fatal)       88 Not Applicable (Non-Fatal)
#> 3        7                          7        1                    January
#> 4       88 Not Applicable (Non-fatal)       88 Not Applicable (Non-Fatal)
#> 5       88 Not Applicable (Non-fatal)       88 Not Applicable (Non-Fatal)
#>   DEATH_TM DEATH_YR               DEATH_YRNAME DOA        DOANAME DRINKING
#> 1     2216     2020                       2020   7  Died at Scene        1
#> 2     8888     8888 Not Applicable (Non-fatal)   0 Not Applicable        0
#> 3     1507     2020                       2020   0 Not Applicable        0
#> 4     8888     8888 Not Applicable (Non-fatal)   0 Not Applicable        0
#> 5     8888     8888 Not Applicable (Non-fatal)   0 Not Applicable        0
#>                DRINKINGNAME DRUGRES1 DRUGRES1NAME DRUGRES2 DRUGRES2NAME
#> 1    Yes (Alcohol Involved)       NA           NA       NA           NA
#> 2 No (Alcohol Not Involved)       NA           NA       NA           NA
#> 3 No (Alcohol Not Involved)       NA           NA       NA           NA
#> 4 No (Alcohol Not Involved)       NA           NA       NA           NA
#> 5 No (Alcohol Not Involved)       NA           NA       NA           NA
#>   DRUGRES3 DRUGRES3NAME DRUGS               DRUGSNAME DRUGTST1 DRUGTST1NAME
#> 1       NA           NA     0 No (drugs not involved)       NA           NA
#> 2       NA           NA     0 No (drugs not involved)       NA           NA
#> 3       NA           NA     0 No (drugs not involved)       NA           NA
#> 4       NA           NA     0 No (drugs not involved)       NA           NA
#> 5       NA           NA     0 No (drugs not involved)       NA           NA
#>   DRUGTST2 DRUGTST2NAME DRUGTST3 DRUGTST3NAME DRUG_DET DRUG_DETNAME DSTATUS
#> 1       NA           NA       NA           NA        8 Not Reported       2
#> 2       NA           NA       NA           NA        8 Not Reported       0
#> 3       NA           NA       NA           NA        8 Not Reported       2
#> 4       NA           NA       NA           NA        8 Not Reported       0
#> 5       NA           NA       NA           NA        8 Not Reported       0
#>      DSTATUSNAME EJECTION   EJECTIONNAME EJ_PATH                  EJ_PATHNAME
#> 1     Test Given        8 Not Applicable       0 Ejection Path Not Applicable
#> 2 Test Not Given        0    Not Ejected       0 Ejection Path Not Applicable
#> 3     Test Given        8 Not Applicable       0 Ejection Path Not Applicable
#> 4 Test Not Given        8 Not Applicable       0 Ejection Path Not Applicable
#> 5 Test Not Given        0    Not Ejected       0 Ejection Path Not Applicable
#>   EMER_USE   EMER_USENAME EXTRICAT                     EXTRICATNAME FIRE_EXP
#> 1     <NA>           <NA>        0 Not Extricated or Not Applicable     <NA>
#> 2        0 Not Applicable        0 Not Extricated or Not Applicable        0
#> 3     <NA>           <NA>        0 Not Extricated or Not Applicable     <NA>
#> 4     <NA>           <NA>        0 Not Extricated or Not Applicable     <NA>
#> 5        0 Not Applicable        0 Not Extricated or Not Applicable        0
#>         FIRE_EXPNAME FUNC_SYS               FUNC_SYSNAME HARM_EV HARM_EVNAME
#> 1               <NA>        3 Principal Arterial - Other       8  Pedestrian
#> 2 No or Not Reported        3 Principal Arterial - Other       8  Pedestrian
#> 3               <NA>        4             Minor Arterial       8  Pedestrian
#> 4               <NA>        4             Minor Arterial       8  Pedestrian
#> 5 No or Not Reported        4             Minor Arterial       8  Pedestrian
#>   HISPANIC                    HISPANICNAME HOSPITAL
#> 1        7                    Non-Hispanic        0
#> 2        0 Not A Fatality (not Applicable)        0
#> 3       99                         Unknown        5
#> 4        0 Not A Fatality (not Applicable)        5
#> 5        0 Not A Fatality (not Applicable)        0
#>                    HOSPITALNAME HOUR        HOURNAME IMPACT1    IMPACT1NAME
#> 1 Not Transported for Treatment   22 10:00pm-10:59pm    <NA>           <NA>
#> 2 Not Transported for Treatment   22 10:00pm-10:59pm      12 12 Clock Point
#> 3                    EMS Ground   14   2:00pm-2:59pm    <NA>           <NA>
#> 4                    EMS Ground   14   2:00pm-2:59pm    <NA>           <NA>
#> 5 Not Transported for Treatment   14   2:00pm-2:59pm      12 12 Clock Point
#>   INJ_SEV                  INJ_SEVNAME LAG_HRS LAG_MINS LOCATION
#> 1       4             Fatal Injury (K)       0        1       11
#> 2       0       No Apparent Injury (O)     999       99        0
#> 3       4             Fatal Injury (K)       1        7        1
#> 4       3 Suspected Serious Injury (A)     999       99        1
#> 5       0       No Apparent Injury (O)     999       99        0
#>                                                LOCATIONNAME MAKE      MAKENAME
#> 1 Not at Intersection - On Roadway, Not in Marked Crosswalk <NA>          <NA>
#> 2                               Occupant of a Motor Vehicle   22       Pontiac
#> 3                     At Intersection - In Marked Crosswalk <NA>          <NA>
#> 4                     At Intersection - In Marked Crosswalk <NA>          <NA>
#> 5                               Occupant of a Motor Vehicle   35 Nissan/Datsun
#>   MAK_MOD MAN_COLL
#> 1      NA        0
#> 2      NA        0
#> 3      NA        0
#> 4      NA        0
#> 5      NA        0
#>                                                                    MAN_COLLNAME
#> 1 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#> 2 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#> 3 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#> 4 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#> 5 The First Harmful Event was Not a Collision with a Motor Vehicle in Transport
#>   MINUTE MINUTENAME MOD_YEAR MOD_YEARNAME MONTH MONTHNAME PER_NO PER_TYP
#> 1     15         15     <NA>         <NA>     1   January      1       5
#> 2     15         15     2002         2002     1   January      1       1
#> 3      0          0     <NA>         <NA>     1   January      1       5
#> 4      0          0     <NA>         <NA>     1   January      2       5
#> 5      0          0     2010         2010     1   January      1       1
#>                              PER_TYPNAME P_SF1 P_SF1NAME P_SF2 P_SF2NAME P_SF3
#> 1                             Pedestrian    NA        NA    NA        NA    NA
#> 2 Driver of a Motor Vehicle In-Transport    NA        NA    NA        NA    NA
#> 3                             Pedestrian    NA        NA    NA        NA    NA
#> 4                             Pedestrian    NA        NA    NA        NA    NA
#> 5 Driver of a Motor Vehicle In-Transport    NA        NA    NA        NA    NA
#>   P_SF3NAME RACE RACENAME REST_MIS                 REST_MISNAME REST_USE
#> 1        NA   NA       NA        8 Not a Motor Vehicle Occupant       96
#> 2        NA   NA       NA        0     No Indication of Mis-Use        3
#> 3        NA   NA       NA        8 Not a Motor Vehicle Occupant       96
#> 4        NA   NA       NA        8 Not a Motor Vehicle Occupant       96
#> 5        NA   NA       NA        0     No Indication of Mis-Use        3
#>                   REST_USENAME ROAD_FNC ROAD_FNCNAME ROLLOVER ROLLOVERNAME
#> 1 Not a Motor Vehicle Occupant       NA           NA     <NA>         <NA>
#> 2   Shoulder and Lap Belt Used       NA           NA        0  No Rollover
#> 3 Not a Motor Vehicle Occupant       NA           NA     <NA>         <NA>
#> 4 Not a Motor Vehicle Occupant       NA           NA     <NA>         <NA>
#> 5   Shoulder and Lap Belt Used       NA           NA        0  No Rollover
#>   RUR_URB RUR_URBNAME SCH_BUS SCH_BUSNAME SEAT_POS                 SEAT_POSNAME
#> 1       2       Urban       0          No        0 Not a Motor Vehicle Occupant
#> 2       2       Urban       0          No       11        Front Seat, Left Side
#> 3       2       Urban       0          No        0 Not a Motor Vehicle Occupant
#> 4       2       Urban       0          No        0 Not a Motor Vehicle Occupant
#> 5       2       Urban       0          No       11        Front Seat, Left Side
#>   SEX SEXNAME SPEC_USE   SPEC_USENAME STATE STATENAME STR_VEH STR_VEHNAME
#> 1   2  Female     <NA>           <NA>    24  Maryland       1          NA
#> 2   2  Female        0 No Special Use    24  Maryland       0          NA
#> 3   1    Male     <NA>           <NA>    24  Maryland       1          NA
#> 4   2  Female     <NA>           <NA>    24  Maryland       1          NA
#> 5   1    Male        0 No Special Use    24  Maryland       0          NA
#>   ST_CASE TOW_VEH       TOW_VEHNAME VEH_NO VE_FORMS WORK_INJ
#> 1  240001    <NA>              <NA>      0        1        9
#> 2  240001       0 No Trailing Units      1        1        8
#> 3  240002    <NA>              <NA>      0        1        0
#> 4  240002    <NA>              <NA>      0        1        8
#> 5  240002       0 No Trailing Units      1        1        8
#>                      WORK_INJNAME
#> 1                         Unknown
#> 2 Not Applicable (not a fatality)
#> 3                              No
#> 4 Not Applicable (not a fatality)
#> 5 Not Applicable (not a fatality)