Get Fatality Analysis Reporting System (FARS) data with the FARS API

This function provides a convenient interface for accessing FARS data or data summaries using a range of criteria. The api parameter allows you to call one of the following functions to access DOT NHTSA’s Crash API:

get_fars_crash_list returns a list of fatal crashes that have occurred in multiple states in one or more years.

get_fars_crash_details returns a details of a fatal crash that has occurred in a state for a single year.

get_fars_crashes a list of fatal crashes by location that have occurred throughout U.S.

get_fars_summary provides a count of injury severity that have occurred throughout U.S. including count of fatalities and crashes.

get_fars_year provides one of 20 FARS data tables for a single year. Supports downloading to a CSV or JSON file.

Both get_fars_crash_list and get_fars_crashes limit the returned data to 5000 records so consider limiting the range of years requested if data exceeds that threshold.

This package also enables access to the FARS data available through the NHTSA data downloads server in a zip format. Set api to "zip" or use the get_fars_zip function to download this data.

Usage

get_fars(
  year = 2019,
  state,
  county = NULL,
  api = c("crashes", "details", "state list", "summary count", "year dataset", "zip"),
  geometry = FALSE,
  crs = 4326,
  type,
  case,
  details = NULL,
  vehicles,
  pr,
  format,
  path = NULL,
  download = FALSE
)

get_fars_crashes(
  year = NULL,
  start_year,
  end_year,
  state,
  county = NULL,
  geometry = FALSE,
  crs = 4326
)

get_fars_crash_details(
  year = 2019,
  state,
  case,
  geometry = FALSE,
  crs = 4326,
  details = NULL
)

get_fars_crash_list(
  year = NULL,
  start_year,
  end_year,
  state,
  vehicles = c(1, 50)
)

get_fars_summary(year = NULL, start_year, end_year, state)

get_fars_year(
  year = 2017,
  type = "accident",
  format = "json",
  path = NULL,
  download = FALSE
)

Arguments

year

numeric vector. Year or range with start and end year. If api is "details", "year dataset", or "zip" (or using the get_fars_crash_details, get_fars_year, or get_fars_zip functions), a single year is required. All other api options support a range with the minimum value is used as a start year and the maximum value used as a end year. Most api options support the years from 2010 to 2019. "year dataset" only supports 2010 to 2017 and "zip" supports 1975 to 2019. start_year and end_year are ignored if year is not NULL.

state

Required. State name, abbreviation, or FIPS number. get_fars_crash_list supports multiple states.

county

County name or FIPS number. Required for get_fars_crashes.

api

character. API function to use. Supported values include "crashes", "details", "state list", "summary count", "year dataset", and "zip". Default: "crashes".

geometry

If TRUE, return sf object. Optional for get_fars_crashes.

crs

Coordinate reference system to return for get_fars_crashes if geometry is TRUE.

type

Name of the dataset or data file to download when using the "year dataset" api or get_fars_year. Supported values include "ACCIDENT", "CEVENT", "DAMAGE", "DISTRACT", "DRIMPAIR", "FACTOR", "MANEUVER", "NMCRASH", "NMIMPAIR", "NMPRIOR", "PARKWORK", "PBTYPE", "PERSON", "SAFETYEQ", "VEHICLE", "VEVENT VINDECODE", "VINDERIVED", "VIOLATION", "VISION", and "VSOE". Lowercase or mixed case values are permitted.

case

Case number. Required if api is "details" or using get_fars_crash_details. Multiple case numbers can be provided.

details

Type of detailed crash data to return (either "events" or "vehicles"). If NULL, events or vehicle data are excluded from data returned by get_fars_crash_details. Optional for get_fars_crash_details. Default: NULL.

vehicles

numeric vector with the minimum and maximum number of vehicles, e.g. c(1, 2) for minimum of 1 vehicle and maximum of 2. Required for get_fars_crash_list.

pr

logical. If TRUE and the api is "zip", download FARS data for Puerto Rico. No Puerto Rico data available for years 1975-1977. Default: FALSE

format

Default "json". "csv" is supported when using the "year dataset" api. "sas" is supporting for the "zip" api.

path

File path used if download is TRUE.

download

logical. If TRUE and the api is "year dataset" or "zip", download the data to a file. Default FALSE.

start_year

Start year for crash reports.

end_year

End year for crash reports.